# frontend/src/docs/api.yaml
openapi: 3.0.0
info:
  title: WheelsTrust API
  description: API documentation for the WheelsTrust platform
  version: 1.0.0
  contact:
    name: WheelsTrust Support
    email: support@wheelstrust.com
servers:
  - url: https://api.wheelstrust.com/v1
    description: Production server
  - url: http://localhost:5000/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ServiceProvider:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439013
        name:
          type: string
          example: AutoCare Express
        description:
          type: string
          example: Full-service auto repair shop specializing in domestic and foreign vehicles...
        specialties:
          type: array
          items:
            type: string
          example: [Oil Change, Brake Service, Tire Replacement]
        services:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: serv1
              name:
                type: string
                example: Oil Change
              description:
                type: string
                example: Full synthetic oil change with filter replacement
              price:
                type: number
                example: 49.99
        location:
          type: object
          properties:
            address:
              type: string
              example: 123 Auto Blvd
            city:
              type: string
              example: San Francisco
            state:
              type: string
              example: CA
            zipCode:
              type: string
              example: 94105
        hours:
          type: object
          properties:
            monday:
              type: string
              example: 8:00 AM - 6:00 PM
            tuesday:
              type: string
              example: 8:00 AM - 6:00 PM
            wednesday:
              type: string
              example: 8:00 AM - 6:00 PM
            thursday:
              type: string
              example: 8:00 AM - 6:00 PM
            friday:
              type: string
              example: 8:00 AM - 6:00 PM
            saturday:
              type: string
              example: 9:00 AM - 5:00 PM
            sunday:
              type: string
              example: Closed
        phone:
          type: string
          example: 555-123-4567
        email:
          type: string
          example: service@autocare.com
        website:
          type: string
          example: https://autocare-express.com
        verified:
          type: boolean
          example: true
        status:
          type: string
          enum: [active, inactive, suspended]
          example: active
        images:
          type: array
          items:
            type: string
          example: [https://example.com/image1.jpg, https://example.com/image2.jpg]
        createdAt:
          type: string
          format: date-time
          example: 2023-05-15T10:30:00Z
...
paths:
...
  /service-providers:
    get:
      summary: Get all service providers
      tags: [Service Providers]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: location
          in: query
          schema:
            type: string
        - name: specialty
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: boolean
        - name: rating
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 5
      responses:
        '200':
          description: List of service providers
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      serviceProviders:
                        type: array
                        items:
                          $ref: '#/components/schemas/ServiceProvider'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                            example: 50
                          pages:
                            type: integer
                            example: 5
                          currentPage:
                            type: integer
                            example: 1
                          limit:
                            type: integer
                            example: 10

    post:
      summary: Create a new service provider
      tags: [Service Providers]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, description, specialties, services, location, hours, phone, email]
              properties:
                name:
                  type: string
                description:
                  type: string
                specialties:
                  type: array
                  items:
                    type: string
                services:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      description:
                        type: string
                      price:
                        type: number
                location:
                  type: object
                  properties:
                    address:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    zipCode:
                      type: string
                hours:
                  type: object
                  properties:
                    monday:
                      type: string
                    tuesday:
                      type: string
                    wednesday:
                      type: string
                    thursday:
                      type: string
                    friday:
                      type: string
                    saturday:
                      type: string
                    sunday:
                      type: string
                phone:
                  type: string
                email:
                  type: string
                  format: email
                website:
                  type: string
                  format: uri
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Service provider created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Service provider profile created successfully
                  data:
                    $ref: '#/components/schemas/ServiceProvider'
...